+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## [Unreleased][]
 
+-   Don’t throw if side effect doesn’t exist when removing
+
 ## [1.0.0][] - 2019-04-30
 
 ### Added
 
 -   Initial implementation
 
-
-[Unreleased]: https://github.com/niksy/manage-side-effects/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/niksy/manage-side-effects/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/niksy/manage-side-effects/tree/v1.0.0